requirements.txt
discord.py
Flask
psycopg2-binary
dateparser==1.1.8
